Fix bugs in memory and multimemory examples

+++ examples/memory.c
@@ -224,7 +224,7 @@ int main(int argc, const char* argv[]) {
 
   // Grow memory.
   printf("Growing memory...\n");
-  uint32_t old_size;
+  uint64_t old_size;
   error = wasmtime_memory_grow(context, &memory, 1, &old_size);
   if (error != NULL)
     exit_with_error("failed to grow memory", error, trap);
